[Fix #4270] Fix false positive in `Performance/RegexpMatch` for named captures
Fix #4270

Problem
=======

If a regular expression is in lhs of `=~` operator, named captures are assigned to local variables.

For example:

```ruby
if /(?<num>\d+)/ =~ 'a10'
  p num # => '10'
end
```

However, `Performance/RegexpMatch` cop is not aware of this assignment.

Solution
=========

This change makes not to add offenses If a regexp has named captures.
For example:

```ruby
 # This cop adds an offense to the code.
if /(?<num>\d+)/ =~ 'a10'
  p num # => '10'
end

 # This cop does not add offenses to the code.
if /\d/ =~ 'a10'
  p num # => NameError
end
```

And this change adds a regexp_node extension.