[Fix #171] Extend auto-correction support for `Performance/Sum`
Fixes #171.

This PR makes `Performance/Sum` cop can be changed auto-correction scope
depending on the value of `SafeAutoCorrect` in .rubocop.yml.

Its auto-correction is marked as safe by default (`SafeAutoCorrect: true`)
to prevent `TypeError` in auto-correced code when initial value is not
specified as shown below:

```ruby
['a', 'b'].sum # => (String can't be coerced into Integer)
```

Therefore if initial value is not specify as in that code, it will not be auto-corrected.

If user always want to enable auto-correction, user can set `SafeAutoCorrect: false`.

```yaml
Performance/Sum:
  SafeAutoCorrect: false
```

`SafeAutoCorrect` remains `true` by default, so the behavior does not change.